aws/requirements.txt
aws-cdk-lib~=2.50.0
aws-cdk.aws-ec2~=1.180.0
constructs>=10.0.0,<11.0.0
aws-cdk.core
